A zomato clone site using HTML and CSS
the clone site on zomato home page. The icon used on the page are fetched through 'www.flaticons.com'. Also the background image and the logo of zomato have been fetch from the original site by inspecting the elements in the Web Browser. It can be done in the following steps:
1.) Open the zomato site home page and right click on any element.
2.) Now, choose inspect option from the appeared menu. 
3.)Now, select the element you want and in the the element tab appeared on the right side of the screen,there will be the code written for the image or the element you selected.